// src/app-public/modules/worker-central/core/models/employerprofile.ts
export class EmployerProfile {
    ID: number;
    UserId: string;
    FirstName: string;
    LastName: string;
    PrimaryEmail: string;
    PrimaryPhoneNo: string;
    PrimaryPhoneExt: string;
    Password: string;
    SecondaryEmail: string;
    SecondaryPhoneNo: string;
    Address1: string;
    Address2: string;
    City: string;
    State: string;
    Zip: number;
    CompanyName: string;
    SkillSet: string;
    ProfilePic: string;
    IsActive: boolean;
    UpdatedDate: Date;
    CreatedDate: Date;
    LastLoggedIn: Date;
    EmployerType: number;
    Website: string;
    IsRecruiter: boolean;
    IsScheduler: boolean;
    IsInterviewer: boolean;
    IsRequestPending: boolean;
    Description: string;
    KeyWords: string[];
    CompanyId: number;
    Logo: string;
    FullName: string;
    SecondaryPhoneCountryExt: string;
    PrimaryPhoneCountryExt: string;
    IsAdministrator: boolean;
    constructor() {
        this.ID=0;
        this.UserId=null;
        this.FirstName=null;
        this.LastName=null;
        this.PrimaryEmail=null;
        this.PrimaryPhoneNo=null;
        this.PrimaryPhoneExt=null;
        this.Password=null;
        this.SecondaryEmail=null;
        this.SecondaryPhoneNo=null;
        this.Address1=null;
        this.Address2=null;
        this.City=null;
        this.State=null;
        this.Zip=0;
        this.CompanyName=null;
        this.SkillSet=null;
        this.ProfilePic=null;
        this.IsActive=false;
        this.UpdatedDate=new Date();
        this.CreatedDate=new Date();
        this.LastLoggedIn=new Date();
        this.EmployerType=0;
        this.Website=null;
        this.IsRecruiter=false;
        this.IsScheduler=false;
        this.IsInterviewer=false;
        this.IsRequestPending=false;
        this.Description=null;
        this.KeyWords=[];
        this.CompanyId=0;
        this.Logo=null;
        this.FullName=null;
        this.SecondaryPhoneCountryExt=null;
        this.PrimaryPhoneCountryExt=null;
        this.IsAdministrator=false;
    }
}